pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastcausal"
version = "0.1.0"
authors = [
  { name = "Kelvin O. Lim", email = "lim.kelvino@gmail.com" },
]
description = "Fast, easy-to-use causal discovery analysis tools"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "tetrad-port>=0.1.0",
    "dgraph_flex>=0.1.11",
    "pandas>=1.5",
    "numpy>=1.21",
    "scikit-learn>=1.0",
    "PyYAML>=6.0",
    "tqdm>=4.60",
    "graphviz>=0.20",
    "click>=8.0",
]

[project.optional-dependencies]
sem = ["semopy>=2.3"]
jupyter = [
    "ipykernel",
    "ipywidgets",
    "ipython",
    "matplotlib",
    "seaborn",
]
batch = [
    "python-docx>=0.8",
    "pillow>=9.0",
]
all = ["fastcausal[sem,jupyter,batch]"]
dev = [
    "fastcausal[all]",
    "pytest>=7.0",
]

[project.scripts]
fastcausal = "fastcausal.cli:main"

[project.urls]
Homepage = "https://github.com/kelvinlim/fastcausal"
Repository = "https://github.com/kelvinlim/fastcausal"
Issues = "https://github.com/kelvinlim/fastcausal/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fastcausal*"]

[tool.setuptools.package-data]
"fastcausal" = ["data/*.csv", "data/*.txt"]
